Update authentication and homepage routes; add blog template html

+++ routes/web.php
@@ -5,10 +5,15 @@ use App\Http\Controllers\PostController;
 use App\Http\Controllers\ProfileController;
 use Illuminate\Support\Facades\Route;
 
-Route::get('/', function () {
-    return view('welcome');
-});
+// Route::get('/', function () {
+//     return view('welcome');
+// })->name('homePage');
 
+// Home routes
+
+Route::get('/', [HomeController::class, 'homepage'])->name('homePage');
+
+Route::get('/home', [HomeController::class, 'index'])->middleware(['auth', 'admin'])->name('home');
 Route::get('/dashboard', function () {
     return view('dashboard');
 })->middleware(['auth', 'verified'])->name('dashboard');
@@ -23,15 +28,10 @@ require __DIR__ . '/auth.php';
 
 // Posts routes
 
-Route::get('/home', [PostController::class, 'index']);
 Route::get('/posts', [PostController::class, 'index'])->name('posts.index');
 Route::get('/posts/create', [PostController::class, 'create'])->name('posts.create');
 Route::post('/posts', [PostController::class, 'store'])->name('posts.store');
 Route::get('/posts/{post}', [PostController::class, 'show'])->name('posts.show');
 Route::get('/posts/{post}/edit', [PostController::class, 'edit'])->name('posts.edit');
 Route::put('/posts/{post}', [PostController::class, 'update'])->name('posts.update');
-Route::delete('/posts/{post}', [PostController::class, 'destroy'])->name('posts.destroy');
-
-// Home routes
-
-Route::get('/admin/dashboard', [HomeController::class, 'index'])->middleware(['auth', 'admin', 'client']);
+Route::delete('/posts/{post}', [PostController::class, 'destroy'])->middleware(['auth', 'admin'])->name('posts.destroy');
